refactor(video): extract control bar css
