+++ src/options/stitch/styles/runtime/video-control-bar.css
@@ -1,0 +1,111 @@
+.aiob-video-control-bar-button {
+  appearance: none;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  align-self: center;
+  flex: 0 0 auto;
+  width: 31px !important;
+  height: 31px !important;
+  min-width: 31px !important;
+  margin: 0 8px;
+  padding: 0;
+  border: 0 !important;
+  background: transparent !important;
+  color: #fff;
+  cursor: pointer;
+  opacity: 0.94;
+  vertical-align: middle;
+}
+
+.aiob-video-control-bar-button--youtube {
+  transform: translateY(0);
+}
+
+.aiob-video-control-bar-button--bilibili {
+  width: 25px !important;
+  height: 25px !important;
+  min-width: 25px !important;
+  margin: 0 6px;
+  transform: translateY(-4px);
+}
+
+.aiob-video-control-bar-button:hover,
+.aiob-video-control-bar-button:focus-visible {
+  opacity: 1;
+  outline: 0;
+}
+
+.aiob-video-control-bar-button__icon {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: 27px;
+  height: 27px;
+  background: center / contain no-repeat;
+  color: currentColor;
+  font: 700 13px/1 system-ui, sans-serif;
+  pointer-events: none;
+}
+
+.aiob-video-control-bar-button--bilibili .aiob-video-control-bar-button__icon {
+  width: 22px;
+  height: 22px;
+}
+
+.aiob-video-control-bar-popover {
+  --aiob-video-control-accent: #8b5cf6;
+  position: fixed;
+  z-index: 2147483647;
+  display: grid;
+  gap: 10px;
+  width: 220px;
+  padding: 12px;
+  border: 1px solid rgba(255, 255, 255, 0.16);
+  border-radius: 12px;
+  background: rgba(15, 17, 28, 0.96);
+  box-shadow: 0 18px 44px rgba(0, 0, 0, 0.36);
+  color: #fff;
+  font: 13px/1.4 system-ui, sans-serif;
+}
+
+.aiob-video-control-bar-popover[hidden] {
+  display: none;
+}
+
+.aiob-video-control-bar-popover__note-input {
+  width: 100%;
+  box-sizing: border-box;
+  min-height: 34px;
+  padding: 7px 9px;
+  border: 1px solid rgba(255, 255, 255, 0.16);
+  border-radius: 8px;
+  background: rgba(255, 255, 255, 0.08);
+  color: #fff;
+  font: 13px/1.35 system-ui, sans-serif;
+}
+
+.aiob-video-control-bar-popover__note-input::placeholder {
+  color: rgba(255, 255, 255, 0.46);
+}
+
+.aiob-video-control-bar-popover__note-input:focus-visible {
+  border-color: var(--aiob-video-control-accent);
+  box-shadow: 0 0 0 2px rgba(139, 92, 246, 0.28);
+  outline: 0;
+}
+
+.aiob-video-control-bar-popover__option {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  min-height: 24px;
+  color: rgba(255, 255, 255, 0.82);
+}
+
+.aiob-video-control-bar-popover__option input {
+  width: 15px;
+  height: 15px;
+  margin: 0;
+  accent-color: var(--aiob-video-control-accent);
+}
